modepolicy: color format be changed when boot [1/1]
PD#SWPL-173182

Problem:
The user switches to 8k resolution on 8k tv and then switches
to 4k tv. Only the colorspace is switched. After restarting,
the colorspace changes.

Solution:
If the hdmimode selected by the user is not supported by the
current TV, it means that the TV has been changed and the
best_outputmode policy is run.

Verify:
S5

Change-Id: Ic6391f8036299c586c499b9cb8e3fa348ad54d92
Signed-off-by: kaifu.hu <kaifu.hu@amlogic.com>